requirements.txt
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cloudpickle==2.0.0
colorama==0.4.4
configparser==5.0.2
cryptography==3.4.8
cycler==0.10.0
cytoolz==0.11.0
dask==2021.9.0
docker-pycreds==0.4.0
easydict==1.9
experimentlogger-jathr==0.0.5
fonttools==4.26.2
fsspec==2021.8.1
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
imagecodecs==2021.8.26
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.2
locket==0.2.1
matplotlib==3.4.3
munkres==1.1.4
networkx==2.6.3
numexpr==2.7.3
numpy==1.21.2
olefile==0.46
packaging==21.0
pandas==1.3.3
partd==1.2.0
pathtools==0.1.2
Pillow==8.3.2
progress==1.6
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pycparser==2.20
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
sentry-sdk==1.3.1
shortuuid==1.0.1
simplejson==3.17.5
six==1.16.0
smmap==4.0.0
subprocess32==3.5.4
threadpoolctl==2.2.0
tifffile==2021.8.30
toolz==0.11.1
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wandb==0.12.1
wincertstore==0.2
